Modify index.html and style.css
index.html was modified temporarily to include readMe on bottom of html syntax
and the reason being is because i didnt want to continue to keep going back and forth
between files. It was removed in the end, style.css was modified to comply with the
restraints of the challenge, and to complete it.